+++ skfem/importers/__init__.py
@@ -1,1 +1,6 @@
-from .meshio import from_meshio
+import warnings
+
+warnings.warn("DeprecationWarning: skfem.importers was renamed to "
+              "skfem.io and is removed in the next major release.")
+
+from ..io import *
