+++ skfem/io/__init__.py
@@ -1,0 +1,1 @@
+from .meshio import from_meshio
