+++ skfem/io/json.py
